@@ -1,3 +1,3 @@
 module github.com/godcong/chronos/v2
 
-go 1.12
+go 1.18
